Extract some structs into separate files
