+++ wallet.go
@@ -1,16 +1,11 @@
 package main
 
 import (
-	"bytes"
 	"crypto/ecdsa"
 	"crypto/elliptic"
 	"crypto/rand"
 	"crypto/sha256"
-	"encoding/gob"
-	"fmt"
-	"io/ioutil"
 	"log"
-	"os"
 
 	"golang.org/x/crypto/ripemd160"
 )
@@ -24,9 +19,12 @@ type Wallet struct {
 	PublicKey  []byte
 }
 
-// Wallets stores a collection of wallets
-type Wallets struct {
-	Wallets map[string]*Wallet
+// NewWallet creates and returns a Wallet
+func NewWallet() *Wallet {
+	private, public := newKeyPair()
+	wallet := Wallet{private, public}
+
+	return &wallet
 }
 
 // GetAddress returns wallet address
@@ -42,14 +40,6 @@ func (w Wallet) GetAddress() []byte {
 	return address
 }
 
-// NewWallet creates and returns a Wallet
-func NewWallet() *Wallet {
-	private, public := newKeyPair()
-	wallet := Wallet{private, public}
-
-	return &wallet
-}
-
 func newKeyPair() (ecdsa.PrivateKey, []byte) {
 	curve := elliptic.P256()
 	private, err := ecdsa.GenerateKey(curve, rand.Reader)
@@ -59,84 +49,6 @@ func newKeyPair() (ecdsa.PrivateKey, []byte) {
 	pubKey := append(private.PublicKey.X.Bytes(), private.PublicKey.Y.Bytes()...)
 
 	return *private, pubKey
-}
-
-// CreateWallet adds a Wallet to Wallets
-func (ws *Wallets) CreateWallet() string {
-	wallet := NewWallet()
-	address := fmt.Sprintf("%s", wallet.GetAddress())
-
-	ws.Wallets[address] = wallet
-
-	return address
-}
-
-// SaveToFile saves wallets to a file
-func (ws Wallets) SaveToFile() {
-	var content bytes.Buffer
-
-	gob.Register(elliptic.P256())
-
-	encoder := gob.NewEncoder(&content)
-	err := encoder.Encode(ws)
-	if err != nil {
-		log.Panic(err)
-	}
-
-	err = ioutil.WriteFile(walletFile, content.Bytes(), 0644)
-	if err != nil {
-		log.Panic(err)
-	}
-}
-
-// LoadFromFile loads wallets from the file
-func (ws *Wallets) LoadFromFile() error {
-	if _, err := os.Stat(walletFile); os.IsNotExist(err) {
-		return err
-	}
-
-	fileContent, err := ioutil.ReadFile(walletFile)
-	if err != nil {
-		log.Panic(err)
-	}
-
-	var wallets Wallets
-	gob.Register(elliptic.P256())
-	decoder := gob.NewDecoder(bytes.NewReader(fileContent))
-	err = decoder.Decode(&wallets)
-	if err != nil {
-		log.Panic(err)
-	}
-
-	ws.Wallets = wallets.Wallets
-
-	return nil
-}
-
-// GetAddresses returns an array of addresses stored in the wallet file
-func (ws *Wallets) GetAddresses() []string {
-	var addresses []string
-
-	for address := range ws.Wallets {
-		addresses = append(addresses, address)
-	}
-
-	return addresses
-}
-
-// GetWallet returns a Wallet by its address
-func (ws Wallets) GetWallet(address string) Wallet {
-	return *ws.Wallets[address]
-}
-
-// NewWallets creates Wallets and fills it from a file if it exists
-func NewWallets() (*Wallets, error) {
-	wallets := Wallets{}
-	wallets.Wallets = make(map[string]*Wallet)
-
-	err := wallets.LoadFromFile()
-
-	return &wallets, err
 }
 
 // HashPubKey hashes public key
